Adds EGL_EXT_unix_poll extension
Instead of trying to be overly general and define an event mechanism
that could work for multiple operating systems and introducing a
NativeEventObject type as with the EGL_INTEL_native_event_objects
extension this extension instead tries to just focus on supporting unix
based operating systems that can use the poll(2) system call.

This fixes a big problem with the previous extension which was that
there was no portable way for EGL to say what specific events (such
as ready to read or ready to write) the driver was interested in, the
driver could only convey a list of file descriptors.