swap_buffers_with_damage: allow implementing with eglSwapBuffers
This clarifies that it would be valid to implement
eglSwapBuffersWithDamage in terms of eglSwapBuffers since the error
conditions that need to be handled are compatible.

Although this wouldn't be very useful in practice it clarifies that
eglSwapBuffersWithDamage is purely to allow an implementation to
optionally pass damage information to a system compositor for
optimization purposes but otherwise it behaves the same as
eglSwapBuffers.