GLX_EXT_buffer_age: clarify guarantees after query
This clarifies the initial definition of the buffer age to state that it
only relates to the content of pixels that continue to pass the pixel
ownership test since the buffers content was originally created.

This removes the overly general statement allowing "external events" to
reset the buffer age since the only asynchronous interactions with the
buffer age that we define are pixel ownership changes which we are more
explicit about.

This clarifies that although drivers are permitted to reset the age in
response to pixel ownership changes that doesn't affect the validity of
any age already queried for the current frame since the definition of
the buffer age already accounts for pixel ownership changes.