Refine buffer age guarantees by defining pixel age
Introduce readers to the concept of the abstract state "pixel age,"
which can not be queried, but is used as a construct to describe what
guarantees are made about a buffer's pixel content between when its age
is queried and the next frame boundary.

Also, rename GLX_BUFFER_AGE_EXT to GLX_BACK_BUFFER_AGE_EXT, and assign
it an enum value.