GLX_EXT_buffer_age: drop over precise age definition
This drops the more precisely defined buffer age description since it
appears to require implementations to update the age in response to
pixels being added to a buffer (e.g. after an expose event) since those
pixels would have an age of 0 and would usually lower "the minimum age
of all the pixels that make up the buffer."

At this point it seems that the overall spec is clear enough without
this paragraph since we do explicitly define how pixel ownership changes
may optionally affect the age later which should be sufficient. The
interaction with pixel ownership changes is a corner case that we
shouldn't overstate in the spec if it makes it more confusing.