EGL_EXT_swap_buffers_with_damage: version 8 updates
This makes some updates based on discussing the extension with Nvidia.
In particular a detailed issue discussing why the extension now uses a
bottom-left origin for the damage rectangles was added; a detailed issue
was added explaining why damage rectangles have an undefined meaning
when presenting to a resized window and finally Ian Stewart and James
Jones have now been listed as contributors to the spec.